renamed helper to internal
