+++ internal/internal.go
@@ -1,4 +1,4 @@
-package helper
+package internal
 
 import (
 	"encoding/json"
